fix(llama-cpp): keep doctor imports lightweight
