+++ extensions/llama-cpp/src/llama-server-assets.ts
@@ -1,0 +1,105 @@
+import path from "node:path";
+import { resolveLlamaCppDataDir } from "./defaults.js";
+
+export const LLAMA_SERVER_RELEASE = "b10357";
+export const LLAMA_SERVER_BUILD = 10_357;
+export const LLAMA_SERVER_COMMIT = "689e227db485c6b33d061555e74034c93a867649";
+
+export type LlamaServerAsset = {
+  platform: NodeJS.Platform;
+  arch: string;
+  backend: "metal" | "cpu";
+  archive: "tar.gz" | "zip";
+  name: string;
+  sha256: string;
+  executable: string;
+};
+
+const LLAMA_SERVER_ASSETS: LlamaServerAsset[] = [
+  {
+    platform: "darwin",
+    arch: "arm64",
+    backend: "metal",
+    archive: "tar.gz",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-macos-arm64.tar.gz`,
+    sha256: "7f464a2d473d53ebb9c1d7d16db1258ec98c371569816491850686e6a4334c52",
+    executable: "llama-server",
+  },
+  {
+    platform: "darwin",
+    arch: "x64",
+    backend: "cpu",
+    archive: "tar.gz",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-macos-x64.tar.gz`,
+    sha256: "8282cf6b30bfab87080044e98b7b78f2896bfc60414926fae6039a89c0fb6ec2",
+    executable: "llama-server",
+  },
+  {
+    platform: "linux",
+    arch: "arm64",
+    backend: "cpu",
+    archive: "tar.gz",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-ubuntu-arm64.tar.gz`,
+    sha256: "0653b6aa14de35920824045bca7e48f98f629f41943d0fefab1e317bbe8e22a8",
+    executable: "llama-server",
+  },
+  {
+    platform: "linux",
+    arch: "x64",
+    backend: "cpu",
+    archive: "tar.gz",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-ubuntu-x64.tar.gz`,
+    sha256: "6b0ba012036e6d727521100158bfcaa73b460fbb31acab2931c9485f347bd16b",
+    executable: "llama-server",
+  },
+  {
+    platform: "win32",
+    arch: "arm64",
+    backend: "cpu",
+    archive: "zip",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-win-cpu-arm64.zip`,
+    sha256: "a0d73be8d3151d9401fa193f1b83c6d7401191b41786e75f0a60184058333e86",
+    executable: "llama-server.exe",
+  },
+  {
+    platform: "win32",
+    arch: "x64",
+    backend: "cpu",
+    archive: "zip",
+    name: `llama-${LLAMA_SERVER_RELEASE}-bin-win-cpu-x64.zip`,
+    sha256: "6c64cc7db679980fcb34bbbe96b2b1df6127e90ccdad2bd6fbe39532fee2fc4e",
+    executable: "llama-server.exe",
+  },
+];
+
+export function selectLlamaServerAsset(
+  platform: NodeJS.Platform = process.platform,
+  arch = process.arch,
+): LlamaServerAsset {
+  const asset = LLAMA_SERVER_ASSETS.find(
+    (candidate) => candidate.platform === platform && candidate.arch === arch,
+  );
+  if (!asset) {
+    throw new Error(
+      `No verified llama-server ${LLAMA_SERVER_RELEASE} build is available for ${platform}/${arch}. Install a compatible llama-server manually, then rerun llama.cpp setup with its absolute path.`,
+    );
+  }
+  return asset;
+}
+
+export function resolveManagedLlamaServerPaths(asset = selectLlamaServerAsset()): {
+  installDir: string;
+  command: string;
+  presetPath: string;
+} {
+  const installDir = path.join(
+    resolveLlamaCppDataDir(),
+    LLAMA_SERVER_RELEASE,
+    `${asset.platform}-${asset.arch}`,
+  );
+  return {
+    installDir,
+    command: path.join(installDir, asset.executable),
+    presetPath: path.join(resolveLlamaCppDataDir(), "models.ini"),
+  };
+}
